Make UsersDao an interface
