+++ application/src/main/java/com/sanction/thunder/dao/DaoModule.java
@@ -3,6 +3,7 @@ package com.sanction.thunder.dao;
 import com.amazonaws.services.dynamodbv2.document.Table;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import com.sanction.thunder.dao.dynamodb.DynamoDbUsersDao;
 import dagger.Module;
 import dagger.Provides;
 
@@ -14,6 +15,6 @@ public class DaoModule {
   @Singleton
   @Provides
   UsersDao provideUsersDao(Table table, ObjectMapper mapper) {
-    return new UsersDao(table, mapper);
+    return new DynamoDbUsersDao(table, mapper);
   }
 }
